initiallize() finished
