+++ code/meta_data.json
@@ -1,1 +1,1 @@
-{"institution": "\u5c0f\u5b98\u5e02\u5c0f\u5b66", "administrator": "admin", "password": "12345678", "student_days": "8", "teacher_days": "30", "status": "1"}
+{"institution": "abc", "administrator": "123", "password": "123", "student_days": "8", "teacher_days": "30", "status": "1"}
